miscTools/autoTaskReviewer/autotask_mcp/pyproject.toml
[project]
name = "autotask-mcp"
version = "1.0.0"
description = "Production-quality Autotask MCP server for automated ticket and project review"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.0",
]

[project.optional-dependencies]
fastmcp = ["fastmcp>=0.1.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "ruff>=0.1.0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools]
py-modules = ["config", "models", "cache", "client", "server", "server_utils", "run_server"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
